+++ test/Sdk.Generator.Tests/ExtensionStartup/ExtensionStartupRunnerGeneratorTests.cs
@@ -48,21 +48,21 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                                     {
                                         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
                                         [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-                                        internal class WorkerExtensionStartupCodeExecutor : WorkerExtensionStartup
+                                        internal class WorkerExtensionStartupCodeExecutor : global::Microsoft.Azure.Functions.Worker.Core.WorkerExtensionStartup
                                         {
                                             /// <summary>
                                             /// Configures the worker to register extension startup services.
                                             /// </summary>
                                             /// <param name="applicationBuilder">The <see cref="IFunctionsWorkerApplicationBuilder"/> to configure.</param>
-                                            public override void Configure(IFunctionsWorkerApplicationBuilder applicationBuilder)
+                                            public override void Configure(global::Microsoft.Azure.Functions.Worker.IFunctionsWorkerApplicationBuilder applicationBuilder)
                                             {
                                                 try
                                                 {
                                                     new global::Microsoft.Azure.Functions.Tests.WorkerExtensionsSample.SampleExtensionStartup().Configure(applicationBuilder);
                                                 }
-                                                catch (Exception ex)
+                                                catch (global::System.Exception ex)
                                                 {
-                                                    Console.Error.WriteLine("Error calling Configure on Microsoft.Azure.Functions.Tests.WorkerExtensionsSample.SampleExtensionStartup instance."+ex.ToString());
+                                                    global::System.Console.Error.WriteLine("Error calling Configure on Microsoft.Azure.Functions.Tests.WorkerExtensionsSample.SampleExtensionStartup instance."+ex.ToString());
                                                 }
                                             }
                                         }
@@ -132,21 +132,21 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                                     {
                                         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
                                         [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-                                        internal class WorkerExtensionStartupCodeExecutor : WorkerExtensionStartup
+                                        internal class WorkerExtensionStartupCodeExecutor : global::Microsoft.Azure.Functions.Worker.Core.WorkerExtensionStartup
                                         {
                                             /// <summary>
                                             /// Configures the worker to register extension startup services.
                                             /// </summary>
                                             /// <param name="applicationBuilder">The <see cref="IFunctionsWorkerApplicationBuilder"/> to configure.</param>
-                                            public override void Configure(IFunctionsWorkerApplicationBuilder applicationBuilder)
+                                            public override void Configure(global::Microsoft.Azure.Functions.Worker.IFunctionsWorkerApplicationBuilder applicationBuilder)
                                             {
                                                 try
                                                 {
                                                     new global::Microsoft.Azure.Functions.Tests.WorkerExtensionsSample.SampleExtensionStartup().Configure(applicationBuilder);
                                                 }
-                                                catch (Exception ex)
+                                                catch (global::System.Exception ex)
                                                 {
-                                                    Console.Error.WriteLine("Error calling Configure on Microsoft.Azure.Functions.Tests.WorkerExtensionsSample.SampleExtensionStartup instance."+ex.ToString());
+                                                    global::System.Console.Error.WriteLine("Error calling Configure on Microsoft.Azure.Functions.Tests.WorkerExtensionsSample.SampleExtensionStartup instance."+ex.ToString());
                                                 }
                                             }
                                         }