+++ src/DotNetWorker.ApplicationInsights/release_notes.md
@@ -1,5 +1,4 @@
 ## What's Changed
 
-### Microsoft.Azure.Functions.Worker.ApplicationInsights 1.3.0
-- Updating `Azure.Identity` to 1.11.0
-- Enabling support for Microsoft Entra authentication for Application Insights
+### Microsoft.Azure.Functions.Worker.ApplicationInsights 1.3.1
+- Updating `Azure.Identity` to 1.11.4